Adding config screen options.

+++ app/src/main/kotlin/com/currand60/karoocolorspeed/data/ConfigData.kt
@@ -7,13 +7,13 @@ data class ConfigData(
     val useBackgroundColors: Boolean,
     val useArrows: Boolean,
     val useTargetSpeed: Boolean,
-    val stoppedValue: Double,
-    val speedPercentLevel1: Double,
-    val speedPercentLevel2: Double,
-    val speedPercentLevel3: Double,
-    val speedPercentLevel4: Double,
-    val speedPercentLevel5: Double,
-    val targetSpeed: Double,
+    val stoppedValue: Int,
+    val speedPercentLevel1: Int,
+    val speedPercentLevel2: Int,
+    val speedPercentLevel3: Int,
+    val speedPercentLevel4: Int,
+    val speedPercentLevel5: Int,
+    val targetSpeed: Int,
 ) {
     companion object {
         /**
@@ -24,13 +24,13 @@ data class ConfigData(
             useBackgroundColors = true,
             useArrows = true,
             useTargetSpeed = false,
-            stoppedValue = 2.0,
-            speedPercentLevel1 = 50.0,
-            speedPercentLevel2 = 65.0,
-            speedPercentLevel3 = 95.0,
-            speedPercentLevel4 = 105.0,
-            speedPercentLevel5 = 110.0,
-            targetSpeed = 0.0,
+            stoppedValue = 2,
+            speedPercentLevel1 = 50,
+            speedPercentLevel2 = 65,
+            speedPercentLevel3 = 95,
+            speedPercentLevel4 = 105,
+            speedPercentLevel5 = 110,
+            targetSpeed = 0,
         )
     }
 }
